Split the code concerning grid systems into dedicated modules
